[fixed] suppressed compiler warning.  [fixed] bug where MISC items were being created and stored without an item type.
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3devel/MMOCoreORB@340 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/objects/player/PlayerImplementation.cpp
@@ -2904,7 +2904,7 @@ void PlayerImplementation::setEntertainerEvent() {
 		return;
 	}
 	
-	server->addEvent(entertainerEvent, performance->getLoopDuration() * 1000);
+	server->addEvent(entertainerEvent, (uint64) performance->getLoopDuration() * 1000);
 }
 
 void PlayerImplementation::setSurveyEvent(string& resourceName) {
